refactor: extract AppState export-review extension (#580, #430)
Sixth slice of #430 (AppState decomposition). Extracts 3 export-review
passthrough methods (cancelPendingExportReview, setExportReviewResolution,
selectExportReviewMatch) into a new
Sources/BugNarrator/AppState+ExportReview.swift as
`extension AppState { ... }`.

Range: AppState.swift lines 1138-1148 (11 lines, 443 bytes, SHA-256
1d27a8f5e2bda2fc03cec115fd770600728df680cce0168ca14e1c94d6e7ad18). All
3 methods are the simplest delegator shape in the #430 series so far —
single-line passthroughs to issueExportController, no do-catch, no
presenter branch, no async, no @Published mutation.

Confirmed via delegator-vs-orchestrator test
(feedback_delegator_vs_orchestrator.md): all 3 pass, only
`confirmPendingExportReview() async` (immediately after at line 1150)
stays in the class body because it has real orchestration logic.

Triple-proof stack applied per the established protocol:
- Endpoint verification (1138 = func header, 1148 = closing `}`, 1149 =
  trailing blank, 1150 = next-method header, 1151+ = orchestrator body)
- SHA-256 gate matched at both spec-review time (against origin) and
  post-extract time (against the new file's extracted body); byte-diff
  against origin/main is silent (exit 0).
- 662 unit tests pass unchanged; docs-sync validation passes.

Extension inherits @MainActor from AppState's class-level annotation per
SE-0316 — NOT re-annotated. New file imports `import Foundation` only
(SimilarIssueResolution is same-module, UUID/String need Foundation).

No dead-import cleanup — codex confirmed AppState.swift still needs
Combine (ObservableObject/@Published) and Foundation (UUID/URL/
TimeInterval).

Pre-build spec reviewed by codex in a fresh session (3 findings folded:
corrected line ranges post-#576 line-shift, post-deletion blank collapse
note, SimilarIssueResolution location documented).
