+++ Sources/BugNarrator/AppState+ExportReview.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+extension AppState {
+    func cancelPendingExportReview() {
+        issueExportController.cancelPendingExportReview()
+    }
+
+    func setExportReviewResolution(_ resolution: SimilarIssueResolution, for issueID: UUID) {
+        issueExportController.setExportReviewResolution(resolution, for: issueID)
+    }
+
+    func selectExportReviewMatch(_ matchID: String, for issueID: UUID) {
+        issueExportController.selectExportReviewMatch(matchID, for: issueID)
+    }
+}
