+++ Sources/BugNarrator/AppState.swift
@@ -1135,18 +1135,6 @@ final class AppState: ObservableObject {
         }
     }
 
-    func cancelPendingExportReview() {
-        issueExportController.cancelPendingExportReview()
-    }
-
-    func setExportReviewResolution(_ resolution: SimilarIssueResolution, for issueID: UUID) {
-        issueExportController.setExportReviewResolution(resolution, for: issueID)
-    }
-
-    func selectExportReviewMatch(_ matchID: String, for issueID: UUID) {
-        issueExportController.selectExportReviewMatch(matchID, for: issueID)
-    }
-
     func confirmPendingExportReview() async {
         guard let pendingExportReview else {
             return
